+++ identity/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.2.3
+
+- Fix [RUSTSEC-2022-0093] by updating `ed25519-dalek` to `2.0`.
+  See [PR 4337]
+
+[RUSTSEC-2022-0093]: https://rustsec.org/advisories/RUSTSEC-2022-0093
+[PR 4337]: https://github.com/libp2p/rust-libp2p/pull/4337
+
+## 0.2.2
+
+- Implement `from_protobuf_encoding` for RSA `Keypair`.
+  See [PR 4193].
+
+[PR 4193]: https://github.com/libp2p/rust-libp2p/pull/4193
+
+## 0.2.1
+
+- Expose `KeyType` for `PublicKey` and `Keypair`.
+  See [PR 4107].
+
+[PR 4107]: https://github.com/libp2p/rust-libp2p/pull/4107
+
 ## 0.2.0
 
 - Raise MSRV to 1.65.